Moved code to an existing filter
